+++ Core/Render/OpenGL/Renderers/Legacy/World/LegacyWorldRenderer.cs
@@ -15,6 +15,9 @@ using Helion.Resources.Archives.Collection;
 using Helion.Resources.Definitions.Decorate.Properties.Enums;
 using Helion.Util;
 using Helion.Util.Configs;
+using Helion.Util.Configs.Components;
+using Helion.Util.Profiling.Timers;
+using Helion.Util.Timing;
 using Helion.World;
 using Helion.World.Entities;
 using Helion.World.Geometry.Sectors;
@@ -54,6 +57,7 @@ public partial class LegacyWorldRenderer : WorldRenderer
     private readonly Stopwatch m_stopwatch = new();
     private readonly OitFrameBuffer m_oitFrameBuffer = new();
     private readonly RenderInfo m_downSizedRenderInfo = new();
+    private readonly RenderProfiler m_renderProfiler;
     private readonly bool m_vanillaRender;
     private Vec2D m_occludeViewPos;
     private bool m_occlude;
@@ -62,6 +66,7 @@ public partial class LegacyWorldRenderer : WorldRenderer
     private bool m_pixelGapCorrection;
     private bool m_downscaleVanillaBuffer;
     private bool m_postProcessingEffects;
+    private bool m_lastUseBsp;
     private int m_lastTicker = -1;
     private Entity? m_viewerEntity;
     private IWorld? m_previousWorld;
@@ -70,9 +75,10 @@ public partial class LegacyWorldRenderer : WorldRenderer
     private PlaneClipFrameBuffer? m_planeClipFrameBuffer;
     private PlaneClipFrameBuffer? m_wallClipFrameBuffer;
 
-    public LegacyWorldRenderer(IConfig config, ArchiveCollection archiveCollection, LegacyGLTextureManager textureManager)
+    public LegacyWorldRenderer(IConfig config, ArchiveCollection archiveCollection, LegacyGLTextureManager textureManager, RenderProfiler renderProfiler)
     {
         m_config = config;
+        m_renderProfiler = renderProfiler;
         m_entityRenderer = new(config, textureManager, archiveCollection);
         m_primitiveRenderer = new();
         m_worldDataManager = new(m_interpolationProgram);
@@ -128,6 +134,11 @@ public partial class LegacyWorldRenderer : WorldRenderer
         m_lastTicker = -1;
         m_pixelGapCorrection = m_config.Render.PixelGapCorrection.Value;
         m_lastTransferHeightsView = TransferHeightView.Middle;
+        m_bspHeuristics = world.GetBspHeuristics();
+        m_smoothedBspTimeUs = -1;
+        m_bspTimeWindow.Clear();
+        m_aboveThresholdCount = 0;
+        m_belowThresholdCount = 0;
 
         m_stopwatch.Stop();
         Log.Info($"Completed level geometry {m_stopwatch.Elapsed}");
@@ -296,13 +307,19 @@ public partial class LegacyWorldRenderer : WorldRenderer
         // If the transfer height view is not the middle then the cached static geometry cannot be used.
         // Render all sectors dynamically instead.
         m_lastRenderStatic = m_renderStatic;
-        m_renderStatic = !m_config.Developer.ForceBsp.Value && renderInfo.TransferHeightView == TransferHeightView.Middle;
+        m_renderStatic = m_config.Render.Mode.Value != AdaptiveRenderMode.Bsp && renderInfo.TransferHeightView == TransferHeightView.Middle && !m_lastUseBsp;
         m_postProcessingEffects = m_config.Render.PostProcessingEffects;
 
-        var renderTickChange = !m_config.Developer.LockRender.Value && NeedsRenderTickChange(world, renderInfo.TransferHeightView);
+        if (renderInfo.TransferHeightView == TransferHeightView.Middle)
+        {
+            m_lastUseBsp = UseBspBasedOnHeuristic(world);
+            m_renderStatic = !m_lastUseBsp;
+        }
+
+        var renderTickChange = !m_config.Developer.Render.Lock.Value && NeedsRenderTickChange(world, renderInfo.TransferHeightView);
         m_lastTransferHeightsView = renderInfo.TransferHeightView;
 
-        if (!m_config.Developer.LockRender.Value)
+        if (!m_config.Developer.Render.Lock.Value)
             Clear(world, renderInfo, renderTickChange);
 
         m_geometryRenderer.SetRenderMode(m_renderStatic ? GeometryRenderMode.Dynamic : GeometryRenderMode.All, renderInfo.TransferHeightView, renderTickChange);
@@ -317,7 +334,7 @@ public partial class LegacyWorldRenderer : WorldRenderer
         m_downscaleVanillaBuffer = m_config.Render.DownScaleVanillaRenderSampleBuffer.Value > 1;
         SetupClipBuffers(framebuffer, dimension, prevDownscale != m_downscaleVanillaBuffer);
 
-        if (!m_config.Developer.LockRender.Value && renderTickChange)
+        if (!m_config.Developer.Render.Lock.Value && renderTickChange)
             m_entityRenderer.Start(renderInfo);
 
         SetOccludePosition(renderInfo.Camera.PositionInterpolated.Double, renderInfo.Camera.YawRadians, renderInfo.Camera.PitchRadians,
@@ -325,16 +342,19 @@ public partial class LegacyWorldRenderer : WorldRenderer
 
         if (renderTickChange)
         {
+            m_renderProfiler.WorldTraversal.Start();
             SetupRenderData(world, renderInfo);
 
             if (m_renderStatic)
                 IterateBlockmap(world);
             else
                 TraverseBsp(world, renderInfo);
+            m_renderProfiler.WorldTraversal.Stop();
         }
 
         PopulatePrimitives(world);
 
+        m_renderProfiler.WorldGeometry.Start();
         m_geometryRenderer.RenderSkies(renderInfo);
         RenderFloodFill(renderInfo);
 
@@ -366,6 +386,7 @@ public partial class LegacyWorldRenderer : WorldRenderer
             m_entityRenderer.RenderOpaque(renderInfo);
             m_primitiveRenderer.RenderAll(renderInfo);
             RenderTransparent(renderInfo, framebuffer);
+            m_renderProfiler.WorldGeometry.Stop();
             return;
         }
 
@@ -414,6 +435,7 @@ public partial class LegacyWorldRenderer : WorldRenderer
 
         m_entityRenderer.RenderOpaque(renderInfo);
         RenderTransparent(renderInfo, framebuffer);
+        m_renderProfiler.WorldGeometry.Stop();
     }
 
     private void RenderFloodFill(RenderInfo renderInfo)
@@ -421,10 +443,12 @@ public partial class LegacyWorldRenderer : WorldRenderer
         // Doom would draw middle textures over flood fill.
         // Setting the factor using PolygonOffset will push them further away in depth so middle textures are closer and render over.
         // Very tiny for reversed z. Flood fill is pushed in world coordinates in the shader.
+        m_renderProfiler.WorldFloodFill.Start();
         GL.Enable(EnableCap.PolygonOffsetFill);
         SetPolygonOffsetFloodFill();
         m_geometryRenderer.RenderPortals(renderInfo);
         GL.Disable(EnableCap.PolygonOffsetFill);
+        m_renderProfiler.WorldFloodFill.Stop();
     }
 
     private static void SetPolygonOffsetFloodFill()
@@ -580,6 +604,7 @@ public partial class LegacyWorldRenderer : WorldRenderer
         if (!hasEntityFuzzData && !hasEntityAlphaData && !hasDynamicAlphaGeometry && !hasStaticAlphaGeometry)
             return;
 
+        m_renderProfiler.WorldTransparent.Start();
         SetPolygonOffsetFloodFill();
         m_oitFrameBuffer.StartRender();
         GL.DepthMask(false);
@@ -678,6 +703,7 @@ public partial class LegacyWorldRenderer : WorldRenderer
             m_entityRenderer.RenderOitFuzzRefractionPass(renderInfo, true);
 
         GL.DepthMask(true);
+        m_renderProfiler.WorldTransparent.Stop();
     }
 
     private void RenderCompositeStyles(IStyleRenderer styleRenderer)