+++ Core/Render/Renderer.cs
@@ -17,13 +17,13 @@ using Helion.Render.OpenGL.Renderers.Legacy.World;
 using Helion.Render.OpenGL.Renderers.Legacy.World.Automap;
 using Helion.Render.OpenGL.Renderers.Legacy.World.Shader;
 using Helion.Render.OpenGL.Shared;
-using Helion.Render.OpenGL.Texture.Fonts;
 using Helion.Render.OpenGL.Texture.Legacy;
 using Helion.Render.OpenGL.Util;
 using Helion.Resources.Archives.Collection;
 using Helion.Util;
 using Helion.Util.Configs;
 using Helion.Util.Configs.Components;
+using Helion.Util.Profiling.Timers;
 using Helion.Util.Timing;
 using Helion.Window;
 using Helion.World;
@@ -32,7 +32,6 @@ using Helion.World.Geometry.Sectors;
 using NLog;
 using OpenTK.Graphics.OpenGL;
 using System;
-using System.Diagnostics.CodeAnalysis;
 using static Helion.Util.Assertion.Assert;
 
 namespace Helion.Render;
@@ -84,7 +83,7 @@ public partial class Renderer : IDisposable
     public IImageDrawInfoProvider DrawInfo => Textures.ImageDrawInfoProvider;
     private bool UseVirtualResolution => (m_config.Window.Virtual.Enable && m_config.Window.Virtual.Dimension.Value.HasPositiveArea);
 
-    public Renderer(IWindow window, IConfig config, ArchiveCollection archiveCollection, FpsTracker fpsTracker)
+    public Renderer(IWindow window, IConfig config, ArchiveCollection archiveCollection, FpsTracker fpsTracker, RenderProfiler renderProfiler)
     {
         Window = window;
         m_config = config;
@@ -95,7 +94,7 @@ public partial class Renderer : IDisposable
         SetShaderVars();
 
         Textures = new LegacyGLTextureManager(config, archiveCollection);
-        m_worldRenderer = new LegacyWorldRenderer(config, archiveCollection, Textures);
+        m_worldRenderer = new LegacyWorldRenderer(config, archiveCollection, Textures, renderProfiler);
         m_hudRenderer = new LegacyHudRenderer(config, Textures, archiveCollection.DataCache);
         m_automapRenderer = new LegacyAutomapRenderer(archiveCollection);
         m_transitionRenderer = new TransitionRenderer(window);
@@ -115,9 +114,6 @@ public partial class Renderer : IDisposable
 
         PrintGLInfo();
         SetGLStates();
-
-        if (m_config.Developer.ForceBsp)
-            Log.Error("Developer.ForceBsp enabled!");
     }
 
     private mat4 CalculateVirtualMvp(GLFramebuffer buffer, Dimension bufferDimension)