+++ Core/World/Geometry/Sectors/SectorPlane.cs
@@ -108,7 +108,5 @@ public sealed class SectorPlane : SectorSoundSource
         LastRenderChangeGametick = gametick;
     }
 
-   
-
     public override string ToString() => $"Id={Id} Z={Z} Face={Facing} Texture={TextureHandle}";
 }