+++ RELEASENOTES.md
@@ -1,6 +1,7 @@
 # 1.1.0.0 (Pre-release)
 
 ## Features:
+- Add adaptive rendering option. Automatically swaps between BSP and static rendering when beneficial. Largely beneficial for slower integrated GPUs.
 - Add Radsuit intensity.
 - Automatic blood color and fuzz blood options.
 - Initial support for MacOS ARM64 builds.